+++ x-pack/plugins/fleet/server/routes/epm/index.ts
@@ -55,7 +55,6 @@ import {
   getListHandler,
   getInstalledListHandler,
   getLimitedListHandler,
-  getFileHandler,
   getInfoHandler,
   getBulkAssetsHandler,
   installPackageFromRegistryHandler,
@@ -70,6 +69,7 @@ import {
   createCustomIntegrationHandler,
   getInputsHandler,
 } from './handlers';
+import { getFileHandler } from './file_handler';
 
 const MAX_FILE_SIZE_BYTES = 104857600; // 100MB
